apps/api/pyproject.toml
[project]
name = "lattice-api"
version = "0.0.0"
description = "Lattice API — FastAPI backend (local + cloud modes)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "anthropic>=0.40",
    "watchdog>=6.0",
    "xxhash>=3.5",
    "markdown-it-py>=3.0",
    "python-frontmatter>=1.1",
    # cloud-only optional deps; installed by default for now,
    # gated at runtime by --mode flag
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pgvector>=0.3",
    "sqlite-vec>=0.1.9",
    "mcp[cli]>=1.27.1",
    "fastembed>=0.8.0",
]

[project.scripts]
lattice-api = "lattice_api.main:cli"
lattice-mcp = "lattice_api.mcp_server:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lattice_api"]
